backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.10.1
pyjwt==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
